feat(integration): replace stress.test.ts with ephemeral-60min-stress.test.ts
Same 60-min, 10-wallet, 10s-tick combined upload+access stress loop as
before, modernized to use the ephemeral-wallet helpers:

  - Funding: single Multicall3 batch tx instead of 10 sequential transfers
    (~10x faster setup; matches the 100w / 1000w pattern).
  - Refund: refundWallets helper handles parallel sweeps + counts failed
    sweeps. Pre-refund nonce drain loop kept verbatim from the old file
    — needed when a failed tick leaves a pending tx in the mempool that
    would reserve balance and underflow the sweep.
  - Suite gating: skipUnlessSuite('1H-stress-devnet-only') ||
    skipUnlessDevnet(). The workflow's prepare step already hard-rejects
    this suite on non-DevNet networks, but the in-test gate makes
    'pnpm test:stress' against a non-DevNet env a clean skip.
  - Logging: same dual-log pattern (/tmp/cdr-stress.log appended via
    fs.appendFileSync alongside console.log) so the workflow's existing
    artifact upload + tail-80 summary still works without changes.
  - Tick-latency summary at end now uses the shared p50/p95/p99/mean
    helpers from _helpers.ts instead of an in-file percentile().

package.json scripts updated: 'test:integration' excludes the new
filename, 'test:stress' runs the new filename. The old file is removed.
